[refactor] map builder 리팩터링 후 적용
